Add Recipe Feature
completed the feature of adding recipes.

After searching for a recipe, the user is now able to add the recipe to his favorites. that includes the recipe name, instructions, ingredients and the recipe thumbnail.

a new route has been added so that the user can see what recipes he has added to his favorites