Additional Feature: 3
Implemented a share function, the user is now able to share a meal. When clicked on the share button, the link of the meal is copied to the clipboard.

for this to work i implemented a view for the meals. User is now able to view meals before saving them. When the user shares a meal, he shares the link to the meal view