Feedback point 3:
Added a view for custom meals aswell, when creating a custom meal, it is automatically saved, the saved counter in the statistics is corrected accordingly.

User also able to access the custom meal view from the saved meals view.

saved meal view now more uniform, since the details for custom meals is in a seperate view.

Feedback 3 finished