Feedback Code (doc)
Removed redundant documentation, fixed the documentation based on the provided feedback, reduced the amount of line doc, added javadoc like documentation for the frontend methods. added a gitignore file for the .idea repo, since we dont want to have that in our repository.

changed the Name of the Application to: Plan & Plate